rename lint
